fix(scans): remove enums folder
