--- packages/scan/src/enums/AttackParamLocation.ts
@@ -1,9 +1,0 @@
-export enum AttackParamLocation {
-  ARTIFICAL_FRAGMENT = 'artifical-fragment',
-  ARTIFICAL_QUERY = 'artifical-query',
-  BODY = 'body',
-  FRAGMENT = 'fragment',
-  HEADER = 'header',
-  PATH = 'path',
-  QUERY = 'query'
-}
