--- packages/scan/src/enums/Discovery.ts
@@ -1,5 +1,0 @@
-export enum Discovery {
-  CRAWLER = 'crawler',
-  ARCHIVE = 'archive',
-  OAS = 'oas'
-}
