--- packages/scan/src/enums/IntegrationType.ts
@@ -1,10 +1,0 @@
-export enum IntegrationType {
-  SLACK = 'slack',
-  GITHUB = 'github',
-  GITLAB = 'gitlab',
-  JIRA = 'jira',
-  TRELLO = 'trello',
-  AZURE = 'azure',
-  MONDAY = 'monday',
-  SERVICENOW = 'sevicenow'
-}
