--- packages/scan/src/enums/IssueCategory.ts
@@ -1,5 +1,0 @@
-export enum IssueCategory {
-  MEDIUM = 'Medium',
-  HIGH = 'High',
-  LOW = 'Low'
-}
