--- packages/scan/src/enums/Module.ts
@@ -1,4 +1,0 @@
-export enum Module {
-  DAST = 'dast',
-  FUZZER = 'fuzzer'
-}
