--- packages/scan/src/enums/ScanStatus.ts
@@ -1,9 +1,0 @@
-export enum ScanStatus {
-  RUNNING = 'running',
-  PENDING = 'pending',
-  STOPPED = 'stopped',
-  FAILED = 'failed',
-  DONE = 'done',
-  SCHEDULED = 'scheduled',
-  QUEUED = 'queued'
-}
