--- packages/scan/src/enums/index.ts
@@ -1,7 +1,0 @@
-export * from './Module';
-export * from './TestType';
-export * from './Discovery';
-export * from './ScanStatus';
-export * from './IssueCategory';
-export * from './IntegrationType';
-export * from './AttackParamLocation';
